+++ npbench/benchmarks/contour_integral/contour_integral.py
@@ -4,8 +4,7 @@ import numpy as np
 
 
 def rng_complex(shape, rng):
-    return (rng.random(shape, dtype=np.float32) +
-            rng.random(shape, dtype=np.float32) * np.complex64(1j))
+    return (rng.random(shape) + rng.random(shape) * 1j)
 
 
 def initialize(NR, NM, slab_per_bc, num_int_pts):
