+++ npbench/benchmarks/contour_integral/contour_integral_dace.py
@@ -8,13 +8,13 @@ NR, NM, slab_per_bc = (dc.symbol(s, dtype=dc.int64)
 
 
 @dc.program
-def contour_integral(Ham: dc.complex64[slab_per_bc + 1, NR, NR],
-                     int_pts: dc.complex64[32], Y: dc.complex64[NR, NM]):
-    P0 = np.zeros((NR, NM), dtype=np.complex64)
-    P1 = np.zeros((NR, NM), dtype=np.complex64)
+def contour_integral(Ham: dc.complex128[slab_per_bc + 1, NR, NR],
+                     int_pts: dc.complex128[32], Y: dc.complex128[NR, NM]):
+    P0 = np.zeros((NR, NM), dtype=np.complex128)
+    P1 = np.zeros((NR, NM), dtype=np.complex128)
     for idx in range(32):
         z = int_pts[idx]
-        Tz = np.zeros((NR, NR), dtype=np.complex64)
+        Tz = np.zeros((NR, NR), dtype=np.complex128)
         for n in range(slab_per_bc + 1):
             zz = np.power(z, slab_per_bc / 2 - n)
             Tz += zz * Ham[n]
