+++ npbench/benchmarks/mandelbrot1/mandelbrot1_dpnp.py
@@ -2,13 +2,13 @@ import dpnp as np
 
 def mandelbrot(xmin, xmax, ymin, ymax, xn, yn, maxiter, horizon=2.0):
 
-    X = np.linspace(xmin, xmax, xn, dtype=np.float32)
-    Y = np.linspace(ymin, ymax, yn, dtype=np.float32)
+    X = np.linspace(xmin, xmax, xn, dtype=np.float64)
+    Y = np.linspace(ymin, ymax, yn, dtype=np.float64)
     X, Y = np.meshgrid(X, Y)
    # C = X + Y[:, None] * 1j
     C = X + Y * 1j
     N = np.zeros(C.shape, dtype=np.int64)
-    Z = np.zeros(C.shape, dtype=np.complex64)
+    Z = np.zeros(C.shape, dtype=np.complex128)
     
     for n in range(maxiter):
         I = np.less(abs(Z), horizon)
