+++ npbench/benchmarks/mandelbrot1/mandelbrot1_numba_np.py
@@ -21,14 +21,14 @@ def linspace(start, stop, num, dtype):
 def mandelbrot(xmin, xmax, ymin, ymax, xn, yn, maxiter, horizon=2.0):
     # Adapted from https://www.ibm.com/developerworks/community/blogs/jfp/...
     #              .../entry/How_To_Compute_Mandelbrodt_Set_Quickly?lang=en
-    # X = np.linspace(xmin, xmax, xn, dtype=np.float32)
-    # Y = np.linspace(ymin, ymax, yn, dtype=np.float32)
-    X = linspace(xmin, xmax, xn, dtype=np.float32)
-    Y = linspace(ymin, ymax, yn, dtype=np.float32)
+    # X = np.linspace(xmin, xmax, xn, dtype=np.float64)
+    # Y = np.linspace(ymin, ymax, yn, dtype=np.float64)
+    X = linspace(xmin, xmax, xn, dtype=np.float64)
+    Y = linspace(ymin, ymax, yn, dtype=np.float64)
     # C = X + Y[:,None]*1j
     C = X + np.reshape(Y, (yn, 1)) * 1j
     N = np.zeros(C.shape, dtype=np.int64)
-    Z = np.zeros(C.shape, dtype=np.complex64)
+    Z = np.zeros(C.shape, dtype=np.complex128)
     for n in range(maxiter):
         # I = np.less(abs(Z), horizon)
         I = np.less(np.absolute(Z), horizon)
