+++ npbench/benchmarks/polybench/gemm/gemm_dace.py
@@ -5,7 +5,7 @@ NI, NJ, NK = (dc.symbol(s, dtype=dc.int64) for s in ('NI', 'NJ', 'NK'))
 
 
 @dc.program
-def kernel(alpha: dc.float32, beta: dc.float32, C: dc.float32[NI, NJ],
-           A: dc.float32[NI, NK], B: dc.float32[NK, NJ]):
+def kernel(alpha: dc.float64, beta: dc.float64, C: dc.float64[NI, NJ],
+           A: dc.float64[NI, NK], B: dc.float64[NK, NJ]):
 
     C[:] = alpha * A @ B + beta * C
